backend/requirements.txt
fastapi==0.104.1
uvicorn==0.29.0
pymongo==4.5.0
motor==3.3.1
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic[email]==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
bcrypt==4.0.1
email-validator==2.0.0
aiofiles==23.2.1
python-dateutil==2.8.2
mongoengine==0.28.1
PyYAML==6.0.1
reportlab==4.0.4
"starlette>=0.37.2,<0.38.0"
openpyxl==3.1.2
pandas==2.2.2